HSEARCH-2792 Rename DslCrossHierarchyTest to ObjectLoadingCrossHierarchyTest
