+++ orm/src/test/java/org/hibernate/search/test/query/objectloading/mixedhierarchy/College.java
@@ -4,10 +4,14 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.query.dsl.mixedhierarchy;
+package org.hibernate.search.test.query.objectloading.mixedhierarchy;
 
+import javax.persistence.DiscriminatorColumn;
+import javax.persistence.DiscriminatorType;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.Inheritance;
+import javax.persistence.InheritanceType;
 
 import org.hibernate.search.annotations.Indexed;
 
@@ -16,15 +20,17 @@ import org.hibernate.search.annotations.Indexed;
  */
 @Entity
 @Indexed
-public class HighSchool extends School {
+@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
+@DiscriminatorColumn(name = "collegeType", discriminatorType = DiscriminatorType.STRING)
+public class College extends EducationalInstitution {
 
 	@Id
 	Long identifier;
 
-	HighSchool() {
+	College() {
 	}
 
-	public HighSchool(long identifier, String name) {
+	public College(long identifier, String name) {
 		super( name );
 		this.identifier = identifier;
 	}
