+++ orm/src/test/java/org/hibernate/search/test/query/objectloading/mixedhierarchy/CommunityCollege.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.query.dsl.mixedhierarchy;
+package org.hibernate.search.test.query.objectloading.mixedhierarchy;
 
 import javax.persistence.Entity;
