+++ orm/src/test/java/org/hibernate/search/test/query/objectloading/mixedhierarchy/HighSchool.java
@@ -4,14 +4,10 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.query.dsl.mixedhierarchy;
+package org.hibernate.search.test.query.objectloading.mixedhierarchy;
 
-import javax.persistence.DiscriminatorColumn;
-import javax.persistence.DiscriminatorType;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
 
 import org.hibernate.search.annotations.Indexed;
 
@@ -20,17 +16,15 @@ import org.hibernate.search.annotations.Indexed;
  */
 @Entity
 @Indexed
-@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
-@DiscriminatorColumn(name = "collegeType", discriminatorType = DiscriminatorType.STRING)
-public class College extends EducationalInstitution {
+public class HighSchool extends School {
 
 	@Id
 	Long identifier;
 
-	College() {
+	HighSchool() {
 	}
 
-	public College(long identifier, String name) {
+	public HighSchool(long identifier, String name) {
 		super( name );
 		this.identifier = identifier;
 	}
